requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
requests>=2.32.0
pytz>=2024.1
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dateutil>=2.9.0
PyYAML>=6.0.1
google-api-python-client>=2.146.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
APScheduler>=3.10.4
google-auth>=2.34.0

